latest update

+++ src/modules/organization/operations/deleteById.ts
@@ -17,11 +17,25 @@ export default async function (id: string) {
     throw new ApiError(httpStatus.BAD_REQUEST, "Organization still has members!")
   }
 
-  await prisma.member.deleteMany({
-    where: { organizationId: id }
+  const result = await prisma.$transaction(async function (tx: any) {
+    const metricsDeleted = await tx.metrics.deleteMany({
+      where: { organizationId: id }
+    })
+
+    const memberDeleted = await tx.member.deleteMany({
+      where: { organizationId: id }
+    })
+
+    if (!memberDeleted || !metricsDeleted) {
+      throw new ApiError(httpStatus.INTERNAL_SERVER_ERROR, "Something went wrong!")
+    }
+
+    return await prisma.organization.delete({
+      where: { id }
+    })
   })
 
-  return await prisma.organization.delete({
-    where: { id }
-  })
+  if (!result) {
+    throw new ApiError(httpStatus.INTERNAL_SERVER_ERROR, "Something went wrong!")
+  }
 }
